Added support for Bluetooth on more BSD systems.
Gammu support Bluetooth on NetBSD was not quite working so I have
rewritten the minimal FreeBSD support to be generic BSD Bluetooth
support since the API is virtually the same.

I also added service searching along the lines of the BlueZ support for
systems that support it (NetBSD-current at the moment though
FreeBSD-current support is on the way). If a device alias
(/etc/bluetooth/hosts) or MAC address is given, it will be contacted
directly otherwise a device inquiry will be made and visible devices
contacted in turn until it succeeds.  If a channel number was given,
that will be used otherwise the RFCOMM channels will be scanned and the
best one picked as per the BlueZ code.