Deal with line breaks in AT+CMGL="ALL" output
With some crazy encodings (like ASCII!) it is possible for line breaks in an SMS
to be translated to real line breaks in the AT+CMGL="ALL" output from the phone.
This means that not all the odd lines start with +CMGL (and some of the even
one do).  ATGEN_ReplyGetMessageList is modified to deal with this situation.

Signed-off-by: Quintin Hill <quintinhill@gmx.net>