Dynamic sleep time in smsd loop
This patch changes the behavior of sleep time defined by loopsleep
variable in smsd config. When is the main loop delayed by some work with
GSM device it is not need to sleep the full loopsleep time. Now it is
sleep called to loopsleep - (receive + send time), but at least to 1
second.