Use CMake automatic variables for detecting type existance
The return value of check_type_size is now tristate, 0 being used in
multi-architecture builts on OS X. This breaks usage of return value as
check for type existence (0 evaluates to false).

Fortunately CMake automatically defines HAVE_* variables which indicate
whether type is available or not and this is only information we
actually need.