SMSD: files: Write MessageReference to the sent SMS
Before, the file backend only printed the MessageReference,
and didn't add it back to the SMS file.
In contrast, the libsql backend does store it.

Without this patch, When processing the SMS with an external application,
The only way to match a sent SMS with its delivery report would be to parse
gammu-smsd logs.

Signed-off-by: Denis Carikli <denis@eukrea.com>