Try to integrate coverage report into AppVeyor
There doesn't seem to be direct way of doing this, so let's hook into
CTest and wrap every test invocation with OpenCppCoverage and submit
results directly to codecov.

It would be better to merge reports and submit them as one, but this is
still good for testing whether this works at all.

Signed-off-by: Michal Čihař <michal@cihar.com>