Build without PostgreSQL on AppVeyor
With it we're really running into DLL hell, which is great to avoid for
now.

Including any native database backend means that we have to ship all
DLLs needed for it as well. This probably could be automated, but the
automation is just not there yet...

Also what would be even more tricky (or even impossible) is building
with MySQL and PostgreSQL support as both link same libraries in
different versions.

Fixes #236

Signed-off-by: Michal Čihař <michal@cihar.com>