Fixed decoding USSD response with some modems
Use usual heuristics we use for strings when phone claims it's not
encoded text. It seems that modems do not agree whether not encoded text
means text as is or text in current modem charset.

This should work well except the corner cases where content of the USSD
is actually something what looks like UCS2 encoded string, but that
shouls not happen much.

Fixes #358

Signed-off-by: Michal Čihař <michal@cihar.com>