Skip garbage reply on AT+CMGL in PDU mode as in text mode
THe D-Link DWM222 4G USB modem sends a garbage line before the reply
to AT+CMGL. This case was handled in text mode but not in PDU mode,
and this device exhibit the bad behavior in PDU mode too.

This change brings garbage skip behavior in PDU mode on par with text mode.